requirements.txt
absl-py==0.11.0
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
protobuf==3.15.2
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
six==1.15.0
Werkzeug==1.0.1
xlrd==2.0.1
